💳🌕 Checkpoint ./commands/say.js:26554501/964 ./watch.json:26554501/2 ./commands/botinfo.js:26554501/2740 ./commands/help.js:26554501/10

+++ commands/say.js
@@ -1,7 +1,8 @@
 exports.run = async(client, message, args) => {
-    if(!message.member.hasPermission("BAN_MEMBERS")) return message.channel.send('You can\'t use that!')
-        if(!message.guild.me.hasPermission("BAN_MEMBERS")) return message.channel.send('kamu tidak memiliki izin.')
-  
+    
+  if (!message.member.permissions.has("MANAGE_MESSAGES"))
+      return message.channel.send("kamu tidak memiliki izin!");
+
   let say = args.join(' ')
   message.channel.send(say)
   message.delete()
